Created a new useSelectableEventLanguages hook from event form code.
PT-981
